+++ FinalProjectTesting/appsettings.json
@@ -1,10 +1,14 @@
 {
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=tcp:grandcircus-2022april.database.windows.net,1433;Initial Catalog=database5;Persist Security Info=False;User ID=Ahbc2022April5@davidluxfordoutlook.onmicrosoft.com;Password=ElectricKoala-Bear96;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Authentication=\"Active Directory Password\""
+  },
+
   "Logging": {
-      "LogLevel": {
+    "LogLevel": {
       "Default": "Information",
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
-      }
-    },
-"AllowedHosts": "*"
+    }
+  },
+  "AllowedHosts": "*"
 }
